prometheus-grafana/grafana-alloy/otel-demo/requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
